Event-Driven Eco Message Retrieval (#44)
I have removed the poll based system for fetching messages and replaced it with a listener on the ChatSent event.
This should finally make us catch 100% of the sent chat messages!

It also gets rid of an annoying hang on server shutdown that our polling thread caused before.

Included in this PR is also some cleanup and an echo command for Discord that can be used to test features without having to keep the Eco client connected.